Revert "Move `GeneratorBlockInfo`."
This reverts commit f077801fc954e74ca3e5eedc2f7264f427e9332b.
